separated transformed message from abstract
